pyproject.toml
[project]
name = "discord-cronbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "discord.py==2.4.0",
  "APScheduler==3.10.4",
  "pydantic-settings==2.5.2",
  "aiosqlite==0.20.0",
]

[tool.uv]
dev-dependencies = ["pytest>=8", "pytest-asyncio>=0.23", "mypy>=1.11", "ruff>=0.6"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
